+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+# UNRELEASED
+
+## Added
+- `events.instrument()` added as a decorator you can use to instrument a single function call.
+
+
 # 0.5.1 - 2019-10-21
 
 ## Changed